+++ src/nes/cpu/opcodes/bcs.rs
@@ -5,7 +5,7 @@ use super::branch::Branch;
 pub struct Bcs { }
 
 impl FromRelative for Bcs {
-    fn from_relative(cpu: &mut Cpu) -> usize {
+    fn from_relative(cpu: &mut Cpu) -> u32 {
         Branch::branch_on_true(cpu, |c| c.SR.is_set(Flags::Carry));
 
         2
